+++ RentMe_Laravel/database/migrations/2022_03_09_225654_create_images_table.php
@@ -18,7 +18,7 @@ class CreateImagesTable extends Migration
 
             
             $table->string('image');
-            $table->integer('apartment_id');
+            $table->unsignedBigInteger('apartment_id');
             $table->foreign('apartment_id')->references('id')->on('apartments');
             
             $table->timestamps();
