+++ RentMe_Laravel/database/migrations/2022_03_09_225921_create_tours_table.php
@@ -22,7 +22,7 @@ class CreateToursTable extends Migration
             $table->string('date');
             $table->string('time');
 
-            $table->integer('apartment_id');
+            $table->unsignedBigInteger('apartment_id');
             $table->foreign('apartment_id')->references('id')->on('apartments');
 
             $table->string('approved');
